tidy OpenCL backend

+++ src/backend/opencl/sort_by_key.cpp
@@ -39,7 +39,7 @@ void sort_by_key(Array<Tk> &okey, Array<Tv> &oval, const Array<Tk> &ikey,
             af::dim4 reorderDims(0, 1, 2, 3);
             reorderDims[dim] = 0;
             preorderDims[0]  = okey.dims()[dim];
-            for (int i = 1; i <= (int)dim; i++) {
+            for (unsigned i = 1; i <= dim; i++) {
                 reorderDims[i - 1] = i;
                 preorderDims[i]    = okey.dims()[i - 1];
             }
